TIKA-378: TikaConfig should notify users if it cannot initialize some parser
Throw TikaExceptions when unable to load parser classes. This is a less likely scenario now that the main parser classes all can be loaded even when the required external dependencies are not present, thus any load problems are more likely caused by invalid configuration instead of deployment decisions.

Note that this change reverts the TIKA-217 solution. This should be fine especially since we no longer try to support Java 1.4.

git-svn-id: https://svn.apache.org/repos/asf/lucene/tika/trunk@911457 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 6ef60c2efd0c2607c400ac8b9e66d605a4ea2267