Fix for TIKA-2613 contributed by ewanmellor.
Change -psm on the Tesseract command line to --psm, with two dashes.
This matches a change in Tesseract 4.0 to remove the one-dash version.
It has been deprecated since Nov 2016.

The Tesseract cset is ee201e1f4.

Also, move the config file (i.e. getOutputType in Tika's terms) so that it
is the last parameter on the command line.  Tesseract logs an error
message (though otherwise doesn't fail) if the config file is not the
last thing on the command line.


Former-commit-id: a761cf7005bc0d19e321c2e5811f5246782c4e1d